+++ index.html
@@ -232,5 +232,7 @@
     <div class="img-upload__message  img-upload__message--loading">Загружаем...</div>
   </template>
 <script src="js/main.js"></script>
+<script type = "module" src="js/hedFunction.js"></script>
+<script type = "module" src="js/pulFunction.js"></script>
 </body>
 </html>
