+++ js/main.js
@@ -1,26 +1,3 @@
-function getRandomNumber(min, max) {
-    return Math.floor(Math.random() * (max - min + 1)) + min;
-  }
-  
-  function generatePhotos() {
+import { generatePhotos } from "./hedFunction";
 
-    const photos = [];
-    for (let i = 0; i < 25; i++) {
-      const photo = {
-        id: i,
-        url: `photos/${i}.jpg`,
-        description: `Dima ${i}`,
-        likes: getRandomNumber(15, 200),
-        comments: getRandomNumber(0, 200)
-      };
-      photos.push(photo);
-    }
-    return photos;
-  }
-  
-  function maxLineString (string, length) {
-    return string.length <= length;
-  }
-  const photos = generatePhotos();
-  console.log(photos[0]);
-  
+console.log(generatePhotos());
