+++ js/pulFunction.js
@@ -1,0 +1,9 @@
+export{getRandomNumber};
+export{maxLineString}
+
+function maxLineString (string, length) {
+    return string.length <= length;
+  }
+  function getRandomNumber(min, max) {
+    return Math.floor(Math.random() * (max - min + 1)) + min;
+  }
