+++ package.json
@@ -1,5 +1,6 @@
 {
   "name": "kekstagram",
+  "type": "module",
   "version": "26.0.0",
   "private": true,
   "description": "Личный проект «Кекстаграм (простой)» от HTML Academy",
@@ -10,6 +11,7 @@
   "bugs": {
     "url": "https://github.com/htmlacademy-univer-javascript-1/2345895-kekstagram-3/issues"
   },
+
   "devDependencies": {
     "browser-sync": "2.27.10",
     "eslint": "8.15.0",
